fix: move auth logic to provider
