+++ package.json
@@ -39,7 +39,6 @@
     "@radix-ui/react-slider": "^1.2.3",
     "@radix-ui/react-slot": "^1.1.1",
     "@t3-oss/env-nextjs": "^0.10.1",
-    "@types/jwt-decode": "^3.1.0",
     "@vercel/analytics": "^1.5.0",
     "@vercel/blob": "^0.27.0",
     "class-variance-authority": "^0.7.1",
@@ -98,7 +97,8 @@
     "tailwindcss-animate": "^1.0.7",
     "ts-jest": "^29.2.5",
     "typescript": "^5.5.3",
-    "vercel": "^39.2.6"
+    "vercel": "^39.2.6",
+    "@types/jwt-decode": "^3.1.0"
   },
   "ct3aMetadata": {
     "initVersion": "7.38.1"
