+++ src/app/home/page.tsx
@@ -1,49 +1,17 @@
-'use client'
-
 import Image from "next/image";
 import SlidingBox from "@/components/SlidingBox/SlidingBox";
 import { useLiff } from "@/providers/liff-provider";
-import { useEffect, useState } from "react";
-import { jwtDecode } from "jwt-decode";
+import { useState, useEffect, createContext, useContext, type ReactNode } from "react";
 
-interface DecodedToken {
-    name: string;
-    picture?: string;
-    sub: string;
-}
+interface UserProfile {
+    displayName: string; 
+    pictureUrl?: string; 
+    userId: string
+  }
 
 export default function HomePage() {
 
-    const { liff } = useLiff();
-    const [userProfile, setUserProfile] = useState<{ displayName: string; pictureUrl?: string; userId: string } | null>(null);
-    const [idToken, setIdToken] = useState<String | null>(null);
-  
-    useEffect(() => {
-        console.log("liff:", liff);
-        if (liff) {
-            if (!liff.isLoggedIn()) {
-                console.log("User not logged in, initiating login...");
-                liff.login();
-            } else {
-                console.log("User is already logged in, decoding ID token...");
-                const token = liff.getIDToken();
-                if (token) {
-                    console.log("ID Token:", token);
-                    setIdToken(token)
-                    try {
-                        const decoded: DecodedToken = jwtDecode(token);
-                        setUserProfile({
-                            displayName: decoded.name,
-                            pictureUrl: decoded.picture,
-                            userId: decoded.sub,
-                        });
-                    } catch (err) {
-                        console.error("Error decoding ID token:", err);
-                    }
-                }
-            }
-        }
-    }, [liff]);     
+    const { liff, liffError, idToken, userProfile } = useLiff();
 
     return (
         <>  
@@ -59,28 +27,6 @@ export default function HomePage() {
                     <h1 className="text-[1.8rem] tracking-[0.1em] font-cooper">Sandbox Space</h1>
                     <div className="font-semibold text-[0.9rem] px-2 tracking-wide">FOR ARTISTS, ENTREPRENEURS, AND BEYOND</div>
                 </div>
-                
-                
-                {/* User Profile Section */}
-                {userProfile ? (
-                    <div className="flex flex-col items-center text-white bg-black/50 p-4 rounded-lg shadow-md">
-                        {userProfile.pictureUrl && (
-                            <img
-                                src={userProfile.pictureUrl}
-                                alt="Profile Picture"
-                                width={80}
-                                height={80}
-                                className="rounded-full border border-white"
-                            />
-                        )}
-                        <p className="mt-2 font-semibold text-lg">{userProfile.displayName}</p>
-                        <p className="text-sm text-gray-300">User ID: {userProfile.userId}</p>
-                    </div>
-                ) : (
-                    <div className="text-white bg-black/50 p-4 rounded-lg shadow-md">
-                        {idToken && <p className="text-sm text-red-400">ID Token: {idToken}</p>}
-                    </div>
-                )}
 
                 {/* Sliding Box */}
                 <SlidingBox />
